requirements.txt
Appium-Python-Client==0.22
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
enum34==1.1.6
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
path.py==8.2.1
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.4
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
selenium==2.53.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
update==0.4.4
wcwidth==0.1.7
